+++ src/assets/styles/components/home/main-product/main-product.scss
@@ -23,7 +23,6 @@
   &__photo {
     max-width: 100%;
     position: absolute;
-    bottom: -50px;
     left: 50%;
     transform: translateX(-50%);
     transition: all 2s ease;
