--- src/scripts/const/PRODUCTNEWS.js
@@ -1,2 +1,0 @@
-import ProductNews from "../model/ProductNews";
-export default new ProductNews();
