--- src/scripts/const/PRODUCTS.js
@@ -1,2 +1,0 @@
-import Products from "../model/Products";
-export default new Products();
