+++ src/scripts/const/home/ALLPRODUCTS.js
@@ -1,0 +1,2 @@
+import AllProducts from "../../model/home/AllProducts";
+export default new AllProducts();
