+++ src/scripts/const/home/BESTPRODUCTS.js
@@ -1,0 +1,2 @@
+import BestProducts from "../../model/home/BestProducts";
+export default new BestProducts();
