+++ src/scripts/const/home/MAINPRODUCTS.js
@@ -1,0 +1,2 @@
+import MainProducts from "../../model/home/MainProducts";
+export default new MainProducts();
