+++ src/scripts/const/home/PRODUCTNEWS.js
@@ -1,0 +1,2 @@
+import ProductNews from "../../model/home/ProductNews";
+export default new ProductNews();
